+++ source/changelog.rst
@@ -2,6 +2,8 @@ Changelog
 =========
 04-29-18
 ~~~~~~~~
+* Feature: Allow users to setup exact match queries
+* Change: Added support for Python3
 * Bugfix: Removed extra calls causing an error in the decrypt process
 
 04-26-18